refactor(units): SI electric current header (#4)
