+++ CMakeLists.txt
@@ -87,6 +87,7 @@ set(
     POSU_UNITS_HPP_FILES
         ${CMAKE_CURRENT_LIST_DIR}/posu/units/base_unit.hpp
         ${CMAKE_CURRENT_LIST_DIR}/posu/units/units.hpp
+        ${CMAKE_CURRENT_LIST_DIR}/posu/units/si/electric_current.hpp
         ${CMAKE_CURRENT_LIST_DIR}/posu/units/si/length.hpp
         ${CMAKE_CURRENT_LIST_DIR}/posu/units/si/mass.hpp
         ${CMAKE_CURRENT_LIST_DIR}/posu/units/si/si.hpp
@@ -95,6 +96,7 @@ set(
 set(
     POSU_UNITS_IPP_FILES
         ${CMAKE_CURRENT_LIST_DIR}/posu/units/ipp/base_unit.ipp
+        ${CMAKE_CURRENT_LIST_DIR}/posu/units/si/ipp/electric_current.ipp
         ${CMAKE_CURRENT_LIST_DIR}/posu/units/si/ipp/length.ipp
         ${CMAKE_CURRENT_LIST_DIR}/posu/units/si/ipp/mass.ipp
         ${CMAKE_CURRENT_LIST_DIR}/posu/units/si/ipp/si.ipp
